#51 - feat : post 삭제 및 comment 생성 기능 리팩토링
