+++ client/src/components/MatPickerModal/MatPickerCreate.tsx
@@ -1,6 +1,6 @@
 import React, { useState, PropsWithChildren } from "react";
 import styled from "styled-components";
-import { addMatPickers } from "../../utils/usePickersAxios";
+import { addMatPickers } from "../../utils/axiosAPI/groups/(임시)PickersAxios";
 
 const ModalContainer = styled.div`
   height: 100%;
@@ -68,7 +68,8 @@ const TabButton = styled.li`
   border: solid 1px #373737;
   border-radius: 20px;
   background-color: ${(props) => props.color || "#adadad"};
-  border: ${(props) => (props.id === "focused" ? "2.5px solid #C65D7B" : "1px solid #505050")};
+  border: ${(props) =>
+    props.id === "focused" ? "2.5px solid #C65D7B" : "1px solid #505050"};
 `;
 
 const ButtonContainer = styled.div`
@@ -95,7 +96,9 @@ const tabs = [
   { index: 3, color: "#023f00" },
 ];
 
-const MatPickerCreate = ({ onClickToggleModal }: PropsWithChildren<ModalDefaultType>) => {
+const MatPickerCreate = ({
+  onClickToggleModal,
+}: PropsWithChildren<ModalDefaultType>) => {
   const [colorValue, setColorValue] = useState("");
   const [nameValue, setNameValue] = useState("");
