+++ client/src/utils/axiosAPI/comments/commentsAxios.ts
@@ -9,23 +9,24 @@ export interface IComments {
   createdat: string;
 }
 
+export const getComments = async () => {
+  const response = await axios.get<IComments[]>(`${url}/comments`);
+  return response.data;
+};
+
 export const createComment = async (
   nickname: string,
   profileimg: string,
   comment: string,
   createdat: string
 ): Promise<void> => {
-  try {
-    await axios.post(`${url}/comments`, {
-      nickname,
-      profileimg,
-      comment,
-      createdat,
-    });
-  } catch (error) {
-    console.error("Error", error);
-    throw error;
-  }
+  const response = await axios.post(`${url}/comments`, {
+    nickname,
+    profileimg,
+    comment,
+    createdat,
+  });
+  return response.data;
 };
 
 export const updateComment = async (
@@ -35,24 +36,16 @@ export const updateComment = async (
   createdat: string,
   id: number
 ): Promise<void> => {
-  try {
-    await axios.patch(`${url}/comments/${id}`, {
-      nickname,
-      profileimg,
-      comment,
-      createdat,
-    });
-  } catch (error) {
-    console.error("Error", error);
-    throw error;
-  }
+  const response = await axios.patch(`${url}/comments/${id}`, {
+    nickname,
+    profileimg,
+    comment,
+    createdat,
+  });
+  return response.data;
 };
 
 export const deleteComment = async (id: number): Promise<void> => {
-  try {
-    await axios.delete(`${url}/comments/${id}`);
-  } catch (error) {
-    console.error("Error", error);
-    throw error;
-  }
+  const response = await axios.delete(`${url}/comments/${id}`);
+  return response.data;
 };
